requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
dash>=2.14.0
plotly>=5.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
